app/requirements.txt
Django==3.0.6
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.11.0
